src build elFinder-2.1-047314f

+++ php/elFinderVolumeDriver.class.php
@@ -6504,8 +6504,9 @@ abstract class elFinderVolumeDriver {
 			$o = '';
 			$r = 1;
 			if (substr(PHP_OS, 0, 3) === 'WIN') {
-				exec('rd /S /Q ' . escapeshellarg($dir), $o, $r);
-				if ($r === 0) {
+				if (!is_link($dir) && is_dir($dir)) {
+					exec('rd /S /Q ' . escapeshellarg($dir), $o, $r);
+				} else {
 					exec('del /F /Q ' . escapeshellarg($dir), $o, $r);
 				}
 			} else {
